Create GPT5-LT-v0.2
This iteration of this developed I wanted to migrate things i considered core from the last prototype to this one. The overall goal was just make it more refined. I want to clarify language next time I get to it, and have a exact setup it shows when outputting its final answers.